+++ src/components/service/serviceChildren.js
@@ -23,14 +23,22 @@ import TableRow from '@mui/material/TableRow';
 import TableSortLabel from '@mui/material/TableSortLabel';
 import Paper from '@mui/material/Paper';
 import Checkbox from '@mui/material/Checkbox';
-import Tooltip from '@mui/material/Tooltip';
 import DeleteIcon from '@mui/icons-material/Delete';
 import { visuallyHidden } from '@mui/utils';
 import DeleteDialog from '../shared/messages/cardDelete';
 import { Trans } from 'react-i18next';
+import { getEnv } from '../../env';
+import Autocomplete from '@mui/material/Autocomplete';
+import axios from 'axios';
+import TextField from '@mui/material/TextField';
 import { useTheme } from '@mui/material/styles';
 import useMediaQuery from '@mui/material/useMediaQuery';
 import Grow from '@mui/material/Grow';
+import AddCircleIcon from '@mui/icons-material/AddCircle';
+import ServiceForm from './serviceForm';
+import Tooltip from '@mui/material/Tooltip';
+
+const env = getEnv();
 
 const DialogRounded = styled(Dialog)(() => ({
   '& .MuiPaper-rounded': {
@@ -48,9 +56,13 @@ const Transition = React.forwardRef(function Transition(props, ref) {
   return <Grow direction="up" ref={ref} {...props} />;
 });
 
-export default function ServiceChildren({ masterTitle, setOpen, status, data, getData, color }) {
+export default function ServiceChildren({ masterTitle, setOpen, status, data, getData, color, tenantName_id }) {
   // DELETE
   const [openDeleteDialog, setOpenDeleteDialog] = React.useState(false);
+
+  // EDIT Paths
+  const [editLevel, setEditLevel] = React.useState(false);
+  const [editData, setEditData] = React.useState({});
 
   const handleClickOpenDeleteDialog = () => {
     setOpenDeleteDialog(true);
@@ -60,7 +72,40 @@ export default function ServiceChildren({ masterTitle, setOpen, status, data, ge
     setOpenDeleteDialog(false);
   };
 
-  const rows = data;
+  const handlePropagation = (e) => {
+    e.stopPropagation();
+  };
+
+  const handleData = (data) => {
+    setEditLevel(true);
+    setEditData(data);
+  };
+  const closeHandleData = () => {
+    setEditLevel(false);
+    getData();
+  };
+
+  const addEdit = (data) => {
+    data.map(
+      (thisElement) =>
+        (thisElement.action = (
+          <Tooltip title={<Trans>service.tooltip.editIcon</Trans>}>
+            <IconButton
+              aria-label="subpath"
+              color="secondary"
+              key={thisElement.id}
+              onClick={() => handleData(thisElement)}
+            >
+              <AddCircleIcon />
+            </IconButton>
+          </Tooltip>
+        ))
+    );
+    return data;
+  };
+
+  const [rows, setRows] = React.useState(data);
+
   const descendingComparator = (a, b, orderBy) => {
     if (b[orderBy] < a[orderBy]) {
       return -1;
@@ -95,7 +140,7 @@ export default function ServiceChildren({ masterTitle, setOpen, status, data, ge
     {
       id: 'id',
       numeric: false,
-      disablePadding: true,
+      disablePadding: false,
       label: 'ID'
     },
     {
@@ -103,6 +148,12 @@ export default function ServiceChildren({ masterTitle, setOpen, status, data, ge
       numeric: true,
       disablePadding: false,
       label: 'Path'
+    },
+    {
+      id: 'action',
+      numeric: false,
+      disablePadding: false,
+      label: ''
     }
   ];
 
@@ -129,7 +180,7 @@ export default function ServiceChildren({ masterTitle, setOpen, status, data, ge
           {headCells.map((headCell, index) => (
             <TableCell
               key={index}
-              align={headCell.numeric ? 'right' : 'left'}
+              align={'left'}
               padding={headCell.disablePadding ? 'none' : 'normal'}
               sortDirection={orderBy === headCell.id ? order : false}
             >
@@ -180,10 +231,10 @@ export default function ServiceChildren({ masterTitle, setOpen, status, data, ge
           </Typography>
         ) : (
           <Typography sx={{ flex: '1 1 100%' }} variant="h6" id="tableTitle" component="div">
-            {data.length > 1 ? (
-              <Trans i18nKey="common.table.counterPlural" values={{ data: data.length }} />
+            {rows.length > 1 ? (
+              <Trans i18nKey="common.table.counterPlural" values={{ data: rows.length }} />
             ) : (
-              <Trans i18nKey="common.table.counterSingle" values={{ data: data.length }} />
+              <Trans i18nKey="common.table.counterSingle" values={{ data: rows.length }} />
             )}
           </Typography>
         )}
@@ -289,8 +340,28 @@ export default function ServiceChildren({ masterTitle, setOpen, status, data, ge
   const isSelected = (name) => selected.indexOf(name) !== -1;
 
   const emptyRows = page > 0 ? Math.max(0, (1 + page) * rowsPerPage - rows.length) : 0;
+  const [pathSelected, setPathSelected] = React.useState(null);
+  const getPaths = (thisPath) => {
+    thisPath === null
+      ? setRows(addEdit(data))
+      : axios
+          .get(env.ANUBIS_API_URL + 'v1/tenants/' + data[0].tenant_id + '/service_paths?name=' + thisPath.path)
+          .then((results) => {
+            setRows(addEdit(results.data));
+          });
+  };
+
+  React.useEffect(() => {
+    data.length > 0 ? getPaths(pathSelected) : '';
+  }, [pathSelected, data]);
+
+  React.useEffect(() => {
+    setRows(addEdit(data));
+  }, [data]);
+
   const theme = useTheme();
   const fullScreen = useMediaQuery(theme.breakpoints.down('sm'));
+
   return (
     <div>
       <IconButton aria-label="path" onClick={handleClickOpen}>
@@ -325,6 +396,28 @@ export default function ServiceChildren({ masterTitle, setOpen, status, data, ge
         </CustomDialogTitle>
         <DialogContent sx={{ minHeight: '400px' }}>
           <Grid container>
+            <Grid item xs={12} sx={{ marginBottom: '25px' }}>
+              <Autocomplete
+                id="sub-path-display"
+                sx={{ width: '100%' }}
+                options={data}
+                value={pathSelected}
+                autoHighlight
+                getOptionLabel={(option) => option.path}
+                onChange={(event, value) => setPathSelected(value)}
+                renderInput={(params) => (
+                  <TextField
+                    {...params}
+                    label={<Trans>service.form.parentPath</Trans>}
+                    variant="outlined"
+                    inputProps={{
+                      ...params.inputProps,
+                      autoComplete: 'new-password'
+                    }}
+                  />
+                )}
+              />
+            </Grid>
             <Grid item xs={12}>
               <Box sx={{ width: '100%' }}>
                 <Paper sx={{ width: '100%', mb: 2 }} elevation={1} square={false}>
@@ -365,10 +458,15 @@ export default function ServiceChildren({ masterTitle, setOpen, status, data, ge
                                     }}
                                   />
                                 </TableCell>
-                                <TableCell component="th" id={labelId} scope="row">
+                                <TableCell component="th" padding="normal" align="left" id={labelId} scope="row">
                                   {row.id}
                                 </TableCell>
-                                <TableCell align="right">{row.path}</TableCell>
+                                <TableCell padding="normal" align="left" scope="row">
+                                  {row.path}
+                                </TableCell>
+                                <TableCell padding="normal" align="left" scope="row" onClick={handlePropagation}>
+                                  {row.action}
+                                </TableCell>
                               </TableRow>
                             );
                           })}
@@ -394,6 +492,28 @@ export default function ServiceChildren({ masterTitle, setOpen, status, data, ge
                     onRowsPerPageChange={handleChangeRowsPerPage}
                   />
                 </Paper>
+                <DialogRounded
+                  open={editLevel}
+                  fullWidth={true}
+                  maxWidth={'xl'}
+                  key={editData.id}
+                  TransitionComponent={Transition}
+                  fullScreen={fullScreen}
+                  onClose={closeHandleData}
+                  aria-labelledby="edit"
+                  aria-describedby="edit"
+                >
+                  <ServiceForm
+                    title={<Trans>service.titles.edit</Trans>}
+                    action={'Sub-service-creation'}
+                    service={editData}
+                    key={editData.id}
+                    getServices={getData}
+                    tenantName_id={tenantName_id}
+                    close={closeHandleData}
+                  />
+                  <DialogActions></DialogActions>
+                </DialogRounded>
                 <DeleteDialog
                   open={openDeleteDialog}
                   onClose={handleCloseDeleteDialog}