+++ src/components/shared/speedDial.js
@@ -10,6 +10,7 @@ import Dialog from '@mui/material/Dialog';
 import DialogActions from '@mui/material/DialogActions';
 import { styled } from '@mui/material/styles';
 import { Trans } from 'react-i18next';
+import AddIcon from '@mui/icons-material/Add';
 import { useTheme } from '@mui/material/styles';
 import useMediaQuery from '@mui/material/useMediaQuery';
 import Grow from '@mui/material/Grow';
@@ -59,7 +60,7 @@ export default function MultifunctionButton({ pageType, setOpen, status, data, g
   };
   const actions = [
     {
-      icon: <EditIcon />,
+      icon: pageType.props.action === 'modify' ? <EditIcon /> : <AddIcon />,
       name:
         pageType.props.action === 'modify' ? (
           <Trans>tenant.tooltip.editIcon</Trans>